Revert some of cleanup in aca4ef3f so that atomic-primops-foreign testsuite can still make use of CounterCommon.hs
